LEAN Exits Gracefully If The Algorithm Adds Invalid Symbol (#13)
* LEAN Exits Gracefully If The Algorithm Adds Invalid Symbol

If the algorithm adds an invalid symbol, e.g. GPBEUR, an exception is raised. However, LEAN doesn't exit because the operation is not done in the main thread, but stops getting data if there are valid symbols.
We added a brokerage error message to exit LEAN.

* Fixes Existing Unit Test

API change from https://github.com/QuantConnect/Lean/pull/7493